+++ doc/whats_new.rst
@@ -353,6 +353,11 @@ API changes summary
      :issue:`8922` by :user:Attractadore
 
 
+    - Deprecate the ``raw_values`` argument of the ``decision_function``
+      method in :class:`sklearn.covariance.EllipticEnvelope` in favor of the
+      estimator parameter ``raw_decision``. ``raw_decision`` now makes
+      ``decision_function`` return the negated Mahalanobis distances.
+
 .. _changes_0_18_1:
 
 Version 0.18.1